Move components outside pages folder, #11
